Clock in an out of services implemented
Created a time entry class and database table to represent an instance
of clocking in and out of a service. Also implemented the UI to allow
simple clock in and out at the time you click (no way to specify
time/date). The UI responds appropriately and updates the overview tab
on clock out. Clocking out updates the service total and client total
balances.