Contextual menus and delete for client and service
Implemented the delete function for both clients and services. Also
implemented contextual action menus for both of those objects. When you
long press a client or service in a list a menu appears above allowing
you to delete or view a single item or delete multiple items. The
contextual menu for services also allows you to clock in.